pyproject.toml
[tool.poetry]
name = "dj-poetry-test"
version = "0.1.0"
description = ""
authors = []
license = ""

[tool.poetry.dependencies]
python = "^3.8, <3.9"
djangorestframework = "^3.12.1"
django = "^3.0"
djangorestframework-simplejwt = "^4.4.0"
django-extensions = "^3.0.9"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
isort = "^5.6.4"
rope = "^0.18.0"
flake8 = "^3.8.4"
factory_boy = "^3.1.0"
faker = "^4.14.0"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 90
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
use_parentheses = true
src_paths = ["backend"]
filter_files = true
known_third_party = ["django", "rest_framework", "faker", "factory"]
